+++ CHANGES.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - New `Cookie` class.
 - Move `ServerRequest::isSecure` to `Request::isSecure`.
+- Adds `Request::getTargetPath`.
 
 ## [0.4.0] - 2019-01-22